feat(ui): show file/commit sidebar in drawer Files tab
The activity and kanban drawers' Files tab rendered DiffView alone,
without the file list and commit list that the standalone PR
browser shows in its left sidebar. Multi-file diffs in a drawer
were not navigable.

Extract PullList's inline diffFilesInline snippet into a new
DiffSidebar component (CommitListSection + grouped file list with
optional 10+ file filter), and render it in PullDetail's Files
tab branch as the left pane of a horizontal split:

  [ files-sidebar  |  files-main          ]
  [ - commits      |  - DiffView          ]
  [ - file list    |                      ]

PullList now uses the same DiffSidebar component (replacing the
inline snippet and dead helper functions), so the standalone PR
browser and the drawer share one source of truth.

The PullList diff-focus border now targets the new
.diff-files-wrap container; one diff-view test selector that
relied on the direct-child .pull-list > .diff-files chain was
loosened to .pull-list > .diff-files-wrap.

E2e regression guards in activity-drawer.spec.ts assert the
sidebar shows the Commits header and the diff file rows in both
the activity drawer and the kanban drawer Files tabs.
